Rely heavily on package.__main__ for test execution.
For proper import role selection of test module imports, direct execution
of modules cannot be supported without requiring that test modules import
their dependencies at a later time. This is not desirable.

Modify module execution to be based on the test.__main__ so that subsequent
arguments are interpreted as the module to run. This has the advantage of not repeating
the __name__ == '__main__' sections littered in every test module.