Migrate away from libexecute dependence and implement the Construction fault.io processor.
libexecute was a test to see how well a simple adaption library would work. The short
conclusion is that the adaption needs to be closer to a programming language.
libexecute is still used by libprobe, and the next change set will likely migrate away
from using it.

The libconstruct.Construction processor manages the construction process of a module
graph. Currently, it is hard coded to specific executables and interfaces, but will likely
become properly parameterized as time progresses. Part of the problem is that a command
needs to be configured by the system/user which is what libexecute did. Moving away from
libexecute means that a replacement is needed.

Disable bin.develop while bin.prepare is in flux.