Improve library and fragment support.
This commit lays the foundation for contexts to determine how linkage occurs.
There was some exploration to allow link parameters to define recommendations,
but the variations meant that contexts would have multiple factors produced for
a single target. Fine for fragments as they would not be user referenced or
installed into systems, but questionable for libraries and executables.