Deprecate libfactor.load() notation.
The import dependency was unnecessary and most of a composite factor should
be raw data and dependencies. libfactor.load()'s class update gave certain
features unnatural access points, and the stack crawl is of questionable
portability. Additionally, this allows the removal of a few unnecessary
classes that will likely lead to libdev.Sources being removed as well.