Add an exit stack to the test object that is controlled by the harness.
This was avoided before, but with better harness definitions the ExitStack
can be managed at the appropriate point in a test's lifecycle.

Notably, it is important that a stack is not processed for exit before
a debugging session is enacted in cases of exceptions while running the
test from development.bin.test. The others care less about when, but
the state in the stack will likely be critical to debugging.