Identify libconstruct as the primary functionality of fault.development.
While much remains independent of libconstruct, the focus of the project
is construction per the number of entry points and lines that refer to it
and their direct or indirect dependence on construction.

The boundaries of libdev are ultimately defined by system.libfactor that
provides access to the (inducted) integrals without depending on
development.library for path resolution.