Refactoring: Implement Mechanism, iFactor, and Build.
This migrates away from per-mechanism parameters helping remove impurities
from the abstraction layer.

system.* builds now rely on imaginary Factors to supply libraries and header directories.
A rather important generalization as much of the implementation was specific to system
factors.

Mechanism provides an explicit class for the managing the adapters extracted from a
Context. "Mechanisms" was renamed to Context in order to clarify its meaning and purpose.

Build is the container for all build parameters excepting the chosen adapter.

Adapter remains class-less, but being an arbitrary collection of data in addition to
the required fields might justify its ambiguity.

A Factor ABC has been created in order to delimit the interfaces required for a build.
This is provided explicitly for generated Factors that are not bound to Python package
modules.

Implement a clean command for removing .fpi directories.

Probes remain in a state of flux.