Finish conversion to use inspect build data rather than delineating during structure time.
Extension modules being the exception as they have to be built ahead of time using a
non-inspect intention.

Primarily, this is about isolating functionality and using the inspect intention
encourages the isolation of xml/fragments build. Unit testing of libhtml transformations
and delineation will become dramatically easier in comparison to the status quo
prior to these patches.