Eliminate Probes in favor of Context Parameters and implement extension architecture.
Primarily for supporting metrics and fragments contexts, the extension architecture allows
tools to be installed to collect measurements about the languages that they support.

Currently, this has caused a freature regression with the HTML formatting pipeline as
the collected measurements are in a different format and are not yet joined into
the fragments documents. The accuracy of the reported coverage should improve with
this change as exact syntax areas are used to identify the counters.