Refactor archive to build a filesystem tree that is likely to be supported by common web servers.
Previously a zip archive was built and served with &.daemon. While this provides a cleaner
pathing interface, there is no way to publish represented factors on hosting services like
github.com. The new form materializes `index.html` files for representing factors and
allocates a `src/` directory for containing source specific information.

Extensions to this directory protocol will likely be used by &.daemon as well.
This essentially orphans &.daemon for the moment as archive removed support for the zip
format entirely.

- Rename &.bin.archive to &.bin.print.
- &.daemon functionality is currently unusable.
- &.html incorporated support (from &.daemon) for rendering indexes.
- Project icons are now presumed to be hyperlinks defined in &.documentation.project.