Add Void exception for signalling empty reads in _read_merged.
Leverage the addition by allowing default `unit_name_delta` of no change,
and a more precise (empty) integration_type list when reading `Render`
for operable tests. Normally, an empty itype list is used to trigger a KeyError
to allow any itype, but in this case the exception is signalling the absence
of a processing instruction for that itype. Consequently, it should be determined
that the itype is inoperable regardless of the `[integration-type]` vector.