Introduce construction-context-mode and refactor metrics.
With intention being refactored into Factor Image Features, intercepts
could no longer be applied as the parameter was no longer an exclusive value.
To compensate, `mode` is introduced as a lower level parameter in order to
maintain the virtual context nature that was being used isolate delineation,
identity, and metrics from normal factor processing.

This expands intercepts to allow the variants' form to be designated by
the selected mode and to remap the mode itself. Ultimately, this allows
for virtual construction contexts to be leveraged under a single location.

Refactor telemetry to work in a parallel with the cache used to support
the metrics build. Prior, all information was stored within the cache directory
assigned to the build of the metrics intention. While it would be possible to
maintain that attribute, prefer to isolate telemetry entirely and provide
the path as a source parameter to any integration operation that includes
a metrics (-m) image feature. Additionally, eliminate the metrics link parameter
originally passed in as part of the intention set as `-m` will suffice.