Initialize the system context directly into integration.
Rename, the default, "vectors" context to "machines" as projects are intended
to represent the machine named after them.

Eliminate execution context initialization in preparation for execute
factors being integrated within the generated system context.
Execution and Construction are being conflated as "System Contexts".

The root system context may be relocated back out in the future depending
on whether or not IID conflicts become problematic. Usually, system is not
depended on directly as it is mostly applications, but it is not
unreasonable to have connections to integration for supporting trivial
factor execution.