Eliminate implementation link and build fault-tool relative to the python machine.
With system references having some support for include directories, the implementation
link can be replaced with a system reference. This allows the C environment to take
its normal form with respect to python interfaces; rather than including through
the python/implementation path, Python headers may be included as they would be
outside of a fault integration context. This is only superior in that expectations
and, presumably,  portability can be maintained.

Relocate fault-tool to be instantiated with the python machine. This reduces
the surface area of system.machines in favor of increasing it in machines.python.
The advantage being that fault-tool may be adjusted within the system context
that it is being bound to rather than being a fault installation global instance.