Implement UTF-8 encoding and decoding with surrogate escape error control by default.
While many alternatives are available, a local solution was preferred in order
to avoid a dependency and for exact control over error conditions. Considering the substantial
investement in Python, a compatible surrogate escape default was preferred.

The terminal devices currently use what's available on the target platform. However,
the libc support for decoding is relatively new (?) and requires a particular locale setting.
Further, macOS does not appear to agree with others on the location of certain headers making
the situation more difficult when compiling xdg-xcb on macOS.