Implement a minor abstraction for the real time and monotonic clocks.
- Single integer timestamps with precision control and datum control.
- Clock constants for Y2K+1 aligning with python/fault.time.
- Datum and precision translation interfaces.

Eventually, this should replace most of python/fault.system.clock leaving a
thin type adapter to clock_view_t.

Parameterization of the integer types is a possible, if not likely, generalization. However,
the 64-bit integers will remain the default regardless of future efforts as it aligns
well with nanosecond precision inside a gregorian cycle.