Drop the project alignment formality with &system.factors.test.
Adjust &system.factors.test to behave more like a &system.factors.map call and allow
tests to be ran from arbitrary factor selectors.

While there is a strong pattern for tests to be included alongside of the software being tested,
it is not reasonable to expect this, or similar, as a universal or required trait at this level
of integration.

&system.factors.test had been anticipating the use of some protocol; originally,
`project.test.*`. However, when considering the case of plugins or interface implementations,
there is a strong possibility that tests may be an isolated project.